Fixed data race in handle acquisition

+++ source/kibble/util/sparse_set.h
@@ -242,7 +242,7 @@ public:
 
     void release(T handle)
     {
-        K_ASSERT_FMT(is_valid(handle), "Cannot release unknown handle %d/%d", unguard(handle), guard_value(handle));
+        K_ASSERT_FMT(is_valid(handle), "Cannot release unknown handle %lu/%lu", size_t(unguard(handle)), size_t(guard_value(handle)));
 
         T unguarded = handle & k_handle_mask;
         T index = sparse[unguarded];
